# move Tensor comparisons back to C