@peterbell10 2020-11-16
When I added the wrapper in #27423, torch_function wasn't a concern and the overhead was pretty minimal.  I think the wrapper now causes `__torch_function__` dispatch to happen twice which presumably caused most of the overhead.

IIUC, this does slightly change the `__torch_function__` behaviour though. Previously you could override `Tensor.__eq__` separately from `Tensor.eq`. Not sure if that's desirable though and it's now consistent with the other dunder methods already defined in C++ with `TypeError_to_NotImplemented`.