@robieta 2020-11-16
> When I added the wrapper in #27423, `__torch_function__` wasn't a concern and the overhead was pretty minimal. I think the wrapper now causes `__torch_function__` dispatch to happen twice which presumably caused most of the overhead.
> 
> IIUC, this does slightly change the `__torch_function__` behaviour though. Previously you could override `Tensor.__eq__` separately from `Tensor.eq`. Not sure if that's desirable though and it's now consistent with the other dunder methods already defined in C++ with `TypeError_to_NotImplemented`.

I think you can still override the two independently. (I agree that it's unclear if that's even a good thing or not...) I ran the following on 1.7: https://colab.research.google.com/gist/robieta/3453df1dd29e4f62d3036d8d26808f92/tensor_override.ipynb

I get the same behavior on this PR: namely that for the last case: for the `Tensor` subclass with both `__eq__` override and `__torch_function__` the `__eq__` override takes precedence.