@robieta 2020-11-16
> Note that in 1.7 `__torch_function__` is called with `func=torch.Tensor.__eq__`. I suspect it's now called with `func=torch.Tensor.eq` because that's the method doing the `__torch_function__` handling.

When does it do that? In my gist (1.7.0) it's called with `func=torch.Tensor.eq`.