@peterbell10 2020-11-16
Try running this code in the notebook:
```python
print(torch.Tensor.eq)
print(torch.Tensor.__eq__)
```

The wrapper has the wrong name, but you can tell from how it's printed that it's really the dunder version:
```
<method 'eq' of 'torch._C._TensorBase' objects>
<function _TensorBase.eq at 0x7f1aa1f4dd08>
```