@robieta 2020-11-16
So it seems like the consensus is that the `foo / __foo__` divergence for comparison ops is an anomaly and we can probably just unify them. Test failures are the reverse of https://github.com/pytorch/pytorch/pull/27423, so probably fine to switch them back?