@hameerabbasi 2020-11-17
> Test failures are the reverse of #27423, so probably fine to switch them back?

Yes, IIUC ONNX output isn't particularly stable, so as long as it's correct, we should be good.